refine(occupancy): odd-column origin consistency, near-range spread guard, cast and perf cleanups

+++ realsense2_camera/src/parameters.cpp
@@ -67,7 +67,7 @@ void BaseRealSenseNode::getParameters()
     _parameters_names.push_back(param_name);
 
     param_name = std::string("occupancy_max_range");
-    _occupancy_max_range = _parameters->setParam<double>(param_name, 2.5);
+    _occupancy_max_range = static_cast<float>(_parameters->setParam<double>(param_name, 2.5));
     _parameters_names.push_back(param_name);
 
     param_name = std::string("linear_accel_cov");
